AimlyCompanyFinder/requirements.txt
pymysql
cryptography
python-dotenv
tavily-python
google-genai